NodeView 기본 구조 구현

+++ AOS/app/src/main/java/boostcamp/and07/mindsync/data/model/ColorRGB.kt
@@ -1,7 +1,7 @@
 package boostcamp.and07.mindsync.data.model
 
 data class ColorRGB(
-    val red: Float,
-    val green: Float,
-    val blue: Float,
+    val red: Int,
+    val green: Int,
+    val blue: Int,
 )
